[refactor] Move boost miner definitions to miner implementation
They need not be in the header, so move them to the implementation.
